feat: Enhance API Key Management and Add Activation Code Features
- Updated main.go to pass database connection to router setup.
- Added new GetLogs endpoint in admin_handlers.go for fetching audit logs.
- Removed redundant handlers.go file and integrated its functionality into existing handlers.
- Improved router.go to serve static files and handle HTML responses.
- Modified config.go for better readability.
- Enhanced database initialization in db.go to include new models.
- Created new domain models for APIKey, Statistics, DailyStats, and AdminAuditLog.
- Implemented activation code handling in activation.go with verification, generation, listing, and revocation functionalities.
- Added new service methods for managing API keys and statistics.
- Updated app.js to reflect new API endpoints and improve dashboard functionality.
- Created generate_apikey.go for generating and storing API keys securely.

+++ webtoapp-key-server/internal/service/statistics.go
@@ -61,13 +61,13 @@ func (s *StatisticsService) RecordActivation(appID string) error {
 		if err == gorm.ErrRecordNotFound {
 			// 创建新的统计记录
 			stats = domain.Statistics{
-				AppID:                 appID,
-				TotalActivations:      1,
+				AppID:                   appID,
+				TotalActivations:        1,
 				SuccessfulVerifications: 0,
-				FailedVerifications:    0,
-				TotalDevices:          1,
-				ActiveCodes:           0,
-				RevokedCodes:          0,
+				FailedVerifications:     0,
+				TotalDevices:            1,
+				ActiveCodes:             0,
+				RevokedCodes:            0,
 			}
 			return s.db.Create(&stats).Error
 		}
@@ -101,14 +101,14 @@ func (s *StatisticsService) RecordVerification(appID string, success bool) error
 // RecordDailyStats 记录日统计
 func (s *StatisticsService) RecordDailyStats(appID string, verificationCount, successCount, failureCount, newDevices, codesGenerated, codesRevoked int) error {
 	dailyStats := &domain.DailyStats{
-		AppID:            appID,
-		Date:             time.Now(),
-		VerificationCount: verificationCount,
-		SuccessCount:      successCount,
-		FailureCount:      failureCount,
-		NewDevices:        newDevices,
-		CodesGenerated:    codesGenerated,
-		CodesRevoked:      codesRevoked,
+		AppID:             appID,
+		Date:              time.Now(),
+		VerificationCount: int64(verificationCount),
+		SuccessCount:      int64(successCount),
+		FailureCount:      int64(failureCount),
+		NewDevices:        int64(newDevices),
+		CodesGenerated:    int64(codesGenerated),
+		CodesRevoked:      int64(codesRevoked),
 	}
 
 	return s.db.Create(dailyStats).Error
